Extract coffee-theia package from example
