+++ example/browser-app/package.json
@@ -17,7 +17,7 @@
     "@theia/textmate-grammars": "^0.7.1",
     "@theia/typescript": "^0.7.1",
     "@theia/messages": "^0.7.1",
-    "modelserver-example": "0.0.1"
+    "@modelserver/dev-example": "0.0.1"
   },
   "devDependencies": {
     "@theia/cli": "^0.7.1"
