+++ example/electron-app/package.json
@@ -18,7 +18,7 @@
     "@theia/typescript": "^0.7.1",
     "@theia/messages": "^0.7.1",
     "@theia/electron": "^0.7.1",
-    "modelserver-example": "0.0.1"
+    "@modelserver/coffee-theia": "0.0.1"
   },
   "devDependencies": {
     "@theia/cli": "^0.7.1"
