--- example/modelsever-example/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../configs/base.tsconfig",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "lib",
-    "baseUrl": "."
-  },
-  "include": [
-    "src"
-  ]
-}
